Change src to use index and main
